Resvoled minor issues and cleaned code with Manuel
